+++ src/main/java/com/ldtteam/structurize/client/BlueprintBlockAccess.java
@@ -1,7 +1,7 @@
 package com.ldtteam.structurize.client;
 
 import com.ldtteam.structurize.blueprints.v1.Blueprint;
-import com.ldtteam.structurize.lib.BlueprintUtils;
+import com.ldtteam.structurize.blueprints.v1.BlueprintUtils;
 import com.ldtteam.structurize.blocks.ModBlocks;
 import com.ldtteam.structurize.util.BlockUtils;
 import net.minecraft.world.Difficulty;
